% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele.check.R
\name{allele.check}
\alias{allele.check}
\title{allele.check}
\usage{
allele.check(summary, dbweights, snp.col = "rsid", effect.allele.col = "a1",
  alt.allele.col = "a2", beta.col = "beta")
}
\arguments{
\item{summary}{summary data frame}

\item{dbweights}{db weights produced by function extract_weights}

\item{snp.col}{snp column in summary}

\item{effect.allele.col}{effect allele column in summary}

\item{alt.allele.col}{alternative allele column in summary}

\item{beta.col}{beta column in summary; if you have odds ration, change it beta by log(odds ratio)}
}
\description{
allele.check
}

