% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_weights.R
\name{extract_weights}
\alias{extract_weights}
\title{extract_weights}
\usage{
extract_weights(db)
}
\arguments{
\item{db}{a chracter vector of db filenames with absolute path to read}
}
\description{
extract_weights
}

